requirements.txt
discord.py==2.3.2
cerebras-cloud-sdk==1.0.0